ci: fix docs release pipeline (#117)
This PR fixes the docs release pipeline but using the `is_beta_tag`
variable as a stringified boolean instead of a boolean.

A manual release pipeline is also added temporarily to release the docs,
once the docs release is made this pipeline will be removed and future
releases of the docs will happen alongside the release of the NPM
package in the existing pipeline.