fix!: use bigint for `addCycles` and `getCyclesBalance` (#208)
Uses BigInt for numbers that are actually supposed to be big integers
according to the API.

Unfortunately, this fix required using a custom JSON parser and
serializer, provided by the
[json-with-bigint](https://www.npmjs.com/package/json-with-bigint)
package. Reimplementing it was not worth it. I've pinned the version of
the package so that it minimizes the risk of introducing any
vulnerabilities with future patches of the package.

Additionally, fixes #193.